Completed random ideas app

+++ routes/ideas.js
@@ -38,16 +38,15 @@ router.get("/", async (req, res) => {
 });
 
 // Get single idea
-router.get('/:id', async (req, res) => {
+router.get("/:id", async (req, res) => {
   try {
     const idea = await Idea.findById(req.params.id);
     res.json({ success: true, data: idea });
   } catch (error) {
     console.log(error);
-    res.status(500).json({ success: false, error: 'Something went wrong' });
+    res.status(500).json({ success: false, error: "Something went wrong" });
   }
 });
-
 
 // Add an idea
 router.post("/", async (req, res) => {
@@ -66,31 +65,57 @@ router.post("/", async (req, res) => {
   }
 });
 
-/// Update idea
+// Update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    // Usernames do not match
+    res
+      .status(403)
+      .json({
+        success: false,
+        error: 'You are not authorized to update this resource',
+      });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
   }
 });
 
-// Delete idea
+/// Delete idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    // Usernames do not match
+    res
+      .status(403)
+      .json({
+        success: false,
+        error: 'You are not authorized to delete this resource',
+      });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
